# backend/gen/http/media/server/encode_decode.go
// Code generated by goa, DO NOT EDIT.
//
// media HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/flaccid/google-keep-clone/backend/design

package server

import (
	"context"
	"net/http"

	media "github.com/flaccid/google-keep-clone/backend/gen/media"
	goahttp "goa.design/goa/v3/http"
)

// EncodeDownloadResponse returns an encoder for responses returned by the
// media download endpoint.
func EncodeDownloadResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]byte)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/octet-stream")
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeDownloadRequest returns a decoder for requests sent to the media
// download endpoint.
func DecodeDownloadRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*media.DownloadPayload, error) {
	return func(r *http.Request) (*media.DownloadPayload, error) {
		var payload *media.DownloadPayload
		var (
			noteID       string
			attachmentID string

			params = mux.Vars(r)
		)
		noteID = params["noteId"]
		attachmentID = params["attachmentId"]
		payload = NewDownloadPayload(noteID, attachmentID)

		return payload, nil
	}
}