# backend/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Google Keep Clone API
    description: A clone of the Google Keep REST API, designed to mirror the official API surface.
    version: "1.0"
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /v1/notes/{noteId}/attachments/{attachmentId}:
        get:
            tags:
                - media
            summary: download media
            description: Downloads an attachment.
            operationId: media#download
            produces:
                - application/octet-stream
            parameters:
                - name: noteId
                  in: path
                  description: The ID of the note.
                  required: true
                  type: string
                - name: attachmentId
                  in: path
                  description: The ID of the attachment.
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
                        format: byte
            schemes:
                - http
...